Replace manual spec tables in API tree (using browser-compat arrays)
This change replaces a number of the last remaining manual
Specifications tables with {{Specifications}} macros calls — for cases
where the article cited multiple BCD features and multiple spec URLs. So
this patch uses browser-compat keys with YAML array values to supply the
data necessary for generating the Browser Compatibility and
Specifications sections.